ai-medical-assistant (1)/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
sentence-transformers==2.2.2
torch>=2.2.0,<2.8.0
pandas==2.1.3
deep-translator==1.11.4
langdetect==1.0.9
langchain-community==0.0.8
langchain-core==0.1.0
langchain-ollama==0.1.0
google-generativeai==0.3.2
reportlab==4.0.7
python-dotenv==1.0.0
